requirements.txt
boto3
apache_airflow
pandas
elasticsearch
faker
psycopg2-binary
pytest
typer
poetry